+++ spec/lib/programme_to_iati_status_spec.rb
@@ -3,40 +3,40 @@ require "rails_helper"
 RSpec.describe "#programme_status_to_iati_status" do
   context "when the user sets the programme status for an activity" do
     it "sets the correct IATI status automatically" do
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("01")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("delivery")
       expect(status).to eq "2"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("02")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("planned")
       expect(status).to eq "1"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("03")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("agreement_in_place")
       expect(status).to eq "1"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("04")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("open_for_applications")
       expect(status).to eq "1"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("05")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("review")
       expect(status).to eq "1"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("06")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("decided")
       expect(status).to eq "1"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("07")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("spend_in_progress")
       expect(status).to eq "2"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("08")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("finalisation")
       expect(status).to eq "3"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("09")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("completed")
       expect(status).to eq "4"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("10")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("stopped")
       expect(status).to eq "5"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("11")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("cancelled")
       expect(status).to eq "5"
 
-      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("12")
+      status = ProgrammeToIatiStatus.new.programme_status_to_iati_status("paused")
       expect(status).to eq "6"
     end
   end