+++ CHANGELOG.md
@@ -418,6 +418,7 @@
 - Add "UK DP Named Contact (NF)" (uk_dp_named_contact) to the importer
 - Add "NF Partner Country DP" (country_delivery_partners) to the importer
 - Add link to the support site in the footer
+- Change programme status field to an enum
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-23...HEAD
 [release-23]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-22...release-23